top부분 삭제(home)에 이동 + 로그인
