+++ src/main/java/com/company/idev/HomeController.java
@@ -3,22 +3,37 @@ package com.company.idev;
 import java.text.DateFormat;
 import java.util.Date;
 import java.util.Locale;
+import java.util.Map;
+
+import javax.servlet.http.HttpSession;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttributes;
+
+import com.company.idev.dto.Member;
+import com.company.idev.mapper.MemberMapper;
 
 /**
  * Handles requests for the application home page.
  */
 @Controller
+@SessionAttributes("member")
 public class HomeController {
 	
 	private static final Logger logger = LoggerFactory.getLogger(HomeController.class);
 	
+	@Autowired
+	MemberMapper mapper;
 	/**
 	 * Simply selects the home view to render by returning its name.
 	 */
@@ -34,6 +49,25 @@ public class HomeController {
 		model.addAttribute("serverTime", formattedDate );
 		
 		return "home";
+	}
+	//·Î±×ÀÎ
+	@GetMapping("/login.do")
+	public String login(@ModelAttribute("success") String success) {
+		return "member/MemberLogin";
+	}
+	@PostMapping("/login.do")
+	public String login(@RequestParam Map<String,String> map, Model model,HttpSession session){
+		Member member=mapper.login(map);
+		String url;
+		if(member != null) {
+			session.setAttribute("member", member);
+			model.addAttribute("member",member);
+			model.addAttribute("success","y");
+			url ="redirect:/";
+		}else {
+			url ="redirect:login.do?success=n";
+		}
+		return url;
 	}
