+++ src/main/java/com/company/idev/cotroller/MemberController.java
@@ -11,15 +11,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.company.idev.dto.Member;
 import com.company.idev.mapper.MemberMapper;
 @Controller
-@RequestMapping(value = "/member")
+@RequestMapping(value = {"/member","/"})
 public class MemberController {
 	private static final Logger logger
 	= LoggerFactory.getLogger(MemberController.class);
@@ -50,24 +48,7 @@ public class MemberController {
 		model.addAttribute("msg", msg);
 		return "member/idCheck";
 	}
-	//·Î±×ÀÎ
-	@GetMapping("/login.do")
-	public String login(@ModelAttribute("success") String success) {
-		return "member/MemberLogin";
-	}
-	@PostMapping("/login.do")
-	public String login(@RequestParam Map<String,String> map, Model model){
-		Member member=mapper.login(map);
-		String url;
-		if(member != null) {
-			model.addAttribute("member",member);
-			model.addAttribute("success","y");
-			url ="includes/top";
-		}else {
-			url ="redirect:login.do?success=n";
-		}
-		return url;
-	}
+
